pyproject.toml
[tool.poetry]
name = "lykos"
version = "1.0.0"
description = "Lykos, the Open-Source Werewolf bot - an IRC bot for playing Werewolf/Mafia games"
authors = ["Fayadd21 <dawoodfayad46@gmail.com>"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "src" },
    { include = "oyoyo" }
]

[tool.poetry.dependencies]
python = "^3.11"
antlr4-python3-runtime = ">=4.8.0,<4.9.0"
ruamel-yaml = ">=0.16.0,<1.0.0"
requests = ">=2.24.0,<3.0.0"
cryptography = ">=38.0.4"

[tool.poetry.scripts]
lykos-bot = "wolfbot:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
ruff = "^0.8.0"
pre-commit = "^3.0.0"
sphinx = "^8.2.3"
sphinx-rtd-theme = "^3.0.2"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
# Ignore common issues that are less critical for basic setup
ignore = [
    "E501",
    "N802",
    "N801",
    "N803",
    "N806",
    "N815",
    "F405",
    "F403",
    "F401",
    "E402",
    "N811",
    "N818",
    "UP007",
    "W191",
    "E701",
    "N999",
    "UP035",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

